+++ src/screens/instructions/instructions.component.tsx
@@ -20,9 +20,10 @@ export default function Instructions() {
         selling, and fighting. The object of the game is to pay off your debt to
         the loan shark. Then, make as much money as you can in a 30 day period.
         It takes one day to travel to a new location. If you deal too heavily in
-        drugs, you might run into the police!! Your main drug stash will be in
-        the Bronx. (It's a nice neighborhood) Put money in the bank to earn
-        interest, but be careful, the loan shark also collects interest.
+        drugs, you might run into the police!! Buy guns to protect yourself from
+        muggers and police dogs. Your main drug stash will be in the Bronx.
+        (It's a nice neighborhood) Put money in the bank to earn interest, but
+        be careful, the loan shark also collects interest.
       </div>
       <div>The Prices of drugs per unit are:</div>
       <PriceBox withDialog={false} priceRange={true} />